Extract openapi_core from openapi_executors
